+++ packages/radix-ui-themes/src/components/radio-cards.props.ts
@@ -5,7 +5,7 @@ import { gridPropDefs } from './grid.props.js';
 const sizes = ['1', '2', '3'] as const;
 const variants = ['surface', 'classic'] as const;
 
-const radioCardGroupRootPropDefs = {
+const radioCardsRootPropDefs = {
   ...asChildPropDef,
   size: { type: 'enum', className: 'rt-r-size', values: sizes, default: '2', responsive: true },
   variant: { type: 'enum', className: 'rt-variant', values: variants, default: 'surface' },
@@ -20,4 +20,4 @@ const radioCardGroupRootPropDefs = {
   gap: PropDef<(typeof gridPropDefs.gap.values)[number]>;
 };
 
-export { radioCardGroupRootPropDefs };
+export { radioCardsRootPropDefs };