+++ packages/radix-ui-themes/changelog.md
@@ -78,8 +78,8 @@
     - Progress bar component that indicates completion of a task.
   - `Radio`
     - Standalone element for building your own layouts with radio inputs.
-  - `RadioCardGroup`
-    - Interactive card component to pick one of the values from the list. Parts:
+  - `RadioCards`
+    - Interactive card components to pick one of the values from the list. Parts:
       - `Root`
       - `Item`
   - `Reset`